Add files via upload
Newproduct.java is a modification of an assignment in CSC 151 using the Scanner class. The original assignment asked me to solicit the input of three integers and find their product. The new program, Newproduct.java, takes that concept and applies them to floating integers. The program asks the user to enter three floating numbers and find their product.